8196884: VS2017 Multiple Type Cast Conversion Compilation Errors
Change the type of symbolic constant badAddressVal and introduce specific casts to fix multiple type cast conversion compilation errors.

+++ src/hotspot/share/utilities/globalDefinitions.hpp
@@ -943,8 +943,8 @@ class JavaCallArguments;
 // Special constants for debugging
 
 const jint     badInt           = -3;                       // generic "bad int" value
-const long     badAddressVal    = -2;                       // generic "bad address" value
-const long     badOopVal        = -1;                       // generic "bad oop" value
+const intptr_t badAddressVal    = -2;                       // generic "bad address" value
+const intptr_t badOopVal        = -1;                       // generic "bad oop" value
 const intptr_t badHeapOopVal    = (intptr_t) CONST64(0x2BAD4B0BBAADBABE); // value used to zap heap after GC
 const int      badStackSegVal   = 0xCA;                     // value used to zap stack segments
 const int      badHandleValue   = 0xBC;                     // value used to zap vm handle area
